backend 2/requirements.txt
flask
pydantic>=2
pypdf
reportlab
pytest
google-generativeai>=0.7.2
